Specify file for git pickaxe when possible
If we can find the git file for a traceback line in the lines changed,
use that for git pickaxe.  This makes things faster as git pickaxe can
be slow.  We still want to look at all files if we can't find the file
in the git logs to catch renames and other edge cases.