Distinguish between lines added and removed
Post process pickaxe results to check if the match was for a full line
and track if the line was added or removed in that change.

A commit showing a line from a stacktrace being removed may just be the
commit that fixed the commit. We generally care more about commits that
add lines from the stacktrace.