Add git file to line even if file not changed in commit range
Running pickaxe (git-log -S) is very slow when running over an entire
repo, so try to run it on a specific file as much as possible.

Although we don't expect any results if the file wasn't modified,
leaving this in for now to see if there any edge cases I can't think of.