Add flag to speed up results
By default we run git pickaxe on the entire repo if we can't find the
stacktrace file in the git repo. This behavior means we have a better
chance of finding a match, but it makes things considerably slower.
So add an option to tradeoff a little bit of accuracy for more speed.