removes advertised Python3 support
This is a painful one but ...

The majority of the code seems to have been written with Python3 in-mind but
the API is written for wsgiref under Python2.7. I encountered all sorts of
issues when trying to use Python3 such that this feature was unusable.

We need a hero to port this to Flask or something more modern but sadly I
cannot be that hero today. The easiest way to fix this it is to drop advertised
support for Python3 until the API can be revised with something more Python3
compatible.